Change meaning of sendDocumentOnValidationFailure
This used to be whether or not to send the operation body as the
signature, but now it's just used to determine whether you'd want to
send it as a separate field in the trace. This is turned off by default,
but users could enable it to understand the operations that are being
sent that are failing validation. This is to avoid users sending
personal information without "opting in"